fix(#39): 리뷰 피드백 반영
- GetProductResponse의 price 필드를 BigDecimal로 변경하여 정밀도 손실 방지
- SearchProductsResponse의 totalElements 계산 로직 수정
- 활성 상품 필터링을 데이터베이스 레벨로 이동
- SearchProductsQuery의 size 검증 로직 통합
- 불필요한 import 제거
- SearchProductsUseCaseTest의 불완전한 테스트 수정
- ProductRepository와 JPA Repository에 새로운 메서드 추가

+++ core/product-core/src/main/java/com/commerce/product/application/usecase/GetProductResponse.java
@@ -6,7 +6,7 @@ import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Getter;
 
-import java.util.ArrayList;
+import java.math.BigDecimal;
 import java.util.List;
 import java.util.stream.Collectors;
 
@@ -42,7 +42,7 @@ public class GetProductResponse {
     public static class OptionDetail {
         private final String id;
         private final String name;
-        private final long price;
+        private final BigDecimal price;
         private final String currency;
         private final List<SkuMappingDetail> skuMappings;
         
@@ -54,7 +54,7 @@ public class GetProductResponse {
             return OptionDetail.builder()
                 .id(option.getId())
                 .name(option.getName())
-                .price(option.getPrice().amount().longValue())
+                .price(option.getPrice().amount())
                 .currency(option.getPrice().currency().name())
                 .skuMappings(mappingDetails)
                 .build();
